requirements.txt
Django>=3.2
djangorestframework>=3.12
psycopg2

flake8>=3.9
